trying to deploy

+++ src/main/java/com/serezk4/server/worker/request/ValidateCoordinatesRequest.java
@@ -1,8 +1,14 @@
 package com.serezk4.server.worker.request;
 
+/**
+ * Validate coordinates request.
+ * @param x - x coordinate
+ * @param y - y coordinate
+ * @param r - r coordinate
+ */
 public record ValidateCoordinatesRequest(
-        float x,
-        float y,
-        float r
+        Double x,
+        Double y,
+        Double r
 ){
 }
