+++ src/main/java/com/serezk4/server/worker/response/ValidateCoordinatesResponse.java
@@ -1,9 +1,17 @@
 package com.serezk4.server.worker.response;
 
+/**
+ * Validate coordinates response.
+ * @param x - x coordinate
+ * @param y - y coordinate
+ * @param r - r coordinate
+ * @param result - result of validation
+ * @param bench - benchmark
+ */
 public record ValidateCoordinatesResponse(
-        float x,
-        float y,
-        float r,
+        Double x,
+        Double y,
+        Double r,
         boolean result,
         long bench
 ){
